Sorting by Name and Rating

+++ app/src/main/java/hu/bme/aut/moviebase/data/Movie_.java
@@ -95,7 +95,7 @@ public class Movie_ implements Parcelable {
     public String description;
 
     @ColumnInfo(name = "rating")
-    public float rating;
+    public int rating;
 
     @ColumnInfo(name = "price")
     public int price;
